+++ rest_email_auth/tests/conftest.py
@@ -3,9 +3,20 @@
 
 import pytest
 
-from rest_framework.test import APIRequestFactory
+from rest_framework.test import APIClient, APIRequestFactory
 
 from rest_email_auth import factories
+
+
+@pytest.fixture
+def api_client():
+    """
+    Fixture to get a client for making API requests.
+
+    Returns:
+        An instance of the APIClient provided by DRF.
+    """
+    return APIClient()
 
 
 @pytest.fixture
@@ -17,6 +28,17 @@ def api_rf():
         An instance of the request factory provided by DRF.
     """
     return APIRequestFactory()
+
+
+@pytest.fixture
+def email_confirmation_factory(db):
+    """
+    Fixture to get the factory used to create email confirmations.
+
+    Returns:
+        The factory used to create ``EmailConfirmation`` instances.
+    """
+    return factories.EmailConfirmationFactory
 
 
 @pytest.fixture