+++ test_settings.py
@@ -12,3 +12,23 @@ INSTALLED_APPS = [
 ]
 
 SECRET_KEY = 'secret'
+
+TEMPLATES = [
+    {
+        'BACKEND': 'django.template.backends.django.DjangoTemplates',
+        'DIRS': [],
+        'APP_DIRS': True,
+        'OPTIONS': {
+            'context_processors': [],
+        },
+    },
+]
+
+
+ROOT_URLCONF = 'test_urls'
+
+
+# Basic settings required for the app
+REST_EMAIL_AUTH = {
+    'EMAIL_VERIFICATION_URL': 'https://example.com/verify?key={key}',
+}