+++ test_urls.py
@@ -1,0 +1,6 @@
+from django.conf.urls import include, url
+
+
+urlpatterns = [
+    url(r'^', include('rest_email_auth.urls')),
+]