+++ CHANGELOG.rst
@@ -1,6 +1,19 @@
 Changelog
 =========
 
+v0.2
+----
+
+Features
+  * #4: Send a verification email after registration.
+  * #5: Add an endpoint for verifying email addresses.
+  * #6: Add an endpoint for resending an email verification.
+  * #7: Add a command for cleaning up expired email confirmations.
+
+Miscellaneous
+  * #14: Email addresses must be unique
+
+
 v0.1
 ----